% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{assign_ExpectedDoublet}
\alias{assign_ExpectedDoublet}
\title{Assigns expected doublet ground variable}
\usage{
assign_ExpectedDoublet(val)
}
\description{
Function returns the expected doublet rate based on 10X data dependent on cell count rather than flat rate.
Increasing cell count in smaller samples.
}
