% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{Assign_ExpectedDoublet_variable}
\alias{Assign_ExpectedDoublet_variable}
\title{Assigns expected doublet ground variable}
\usage{
Assign_ExpectedDoublet_variable(val)
}
\description{
Function returns the expected doublet rate based on 10X data dependent on cell count rather than flat rate. Increasing cell count in smaller samples.
Assign_ExpectedDoublet_variable()
}
