% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSI_functions.R
\name{LSI_function}
\alias{LSI_function}
\title{Function to run LSI on a Seurat object.}
\usage{
LSI_function(mat, nComponents, binarize = FALSE)
}
\arguments{
\item{mat}{raw count matrix from Seurat object.}

\item{nComponents}{number of right singular vectors to estimate.}

\item{binarize}{whether to binarize dgCMatrix class.}
}
\value{
list containing LSI results.
}
\description{
Function to run LSI on a Seurat object.
}
