% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{preprocess_AmbientRNA}
\alias{preprocess_AmbientRNA}
\title{Removes ambient RNA from cells using DecontX pipeline}
\usage{
preprocess_AmbientRNA(seurat_list)
}
\arguments{
\item{seurat_list}{List of Seurat objects}
}
\description{
Removes ambient RNA from cells using DecontX pipeline
}
