% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{preprocess_DoubletRemoval}
\alias{preprocess_DoubletRemoval}
\title{Doublet Removal function}
\usage{
preprocess_DoubletRemoval(seurat_list)
}
\arguments{
\item{seurat_list}{List of Seurat objects}
}
\description{
Fully runs the doublet removal protocol through DoubletFinder() based on variable from assign_ExpectedDoublet().
Returns Seurat object with TRUE/FALSE assignment to duplicate cells in META data.
}
