% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{read_folder}
\alias{read_folder}
\title{Variable gene function selection enabling.}
\usage{
read_folder(folder_path)
}
\value{
List of seurat objects
}
\description{
Reads folder, or folder of folder or 10X files to create Seurat objects
}
