% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{remove_GenesFromSeuratList}
\alias{remove_GenesFromSeuratList}
\title{Function to completely remove genes from a list of Seurat objects}
\usage{
remove_GenesFromSeuratList(seurat_list, genes_to_remove, send_message = NULL)
}
\arguments{
\item{seurat_list}{List of Seurat objects to filter.}

\item{genes_to_remove}{vector of gene names to remove completely. If "remove_ensembl" is passed, will automatically detect ENSG genes.}

\item{send_message}{function to send messages to the UI (optional).}
}
\value{
filtered list of Seurat objects.
}
\description{
Function to completely remove genes from a list of Seurat objects
}
