% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{Remove_lowRNA}
\alias{Remove_lowRNA}
\title{Ambient RNA removal from Seurat object}
\usage{
Remove_lowRNA(seurat_obj)
}
\description{
Uses DecontX protocol to approximate ambient RNA then runs this function to remove ambient RNA from counts matrix.
}
