% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSI_functions.R
\name{run_SparseLogX}
\alias{run_SparseLogX}
\title{Sparse Log of X function. Log normalization on sparse matrix, returns sparse matrix.}
\usage{
run_SparseLogX(spmat, logtype = "log2", scale = FALSE, scaleFactor = 10^4)
}
\arguments{
\item{spmat}{raw count matrix from Seurat object in sparse format.}

\item{logtype}{c("log", "log2", "log10"), standardized to log2.}

\item{scaleFactor}{scaling factor, standardized to 10^4.}
}
\value{
Logarithmic sparse matrix
}
\description{
Work around for staying with sparse matrix format in Seurat object to reduce object size.
}
