% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{runDecontX}
\alias{runDecontX}
\title{Ambient RNA removal from Seurat object}
\usage{
runDecontX(seurat_obj, seed = 1)
}
\description{
Fully runs the ambient RNA removal protocol through DecontX.
}
