% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSI_functions.R
\name{runLSI}
\alias{runLSI}
\title{TF-IDF LSI adapted from Jeff Granja, who adapted from flyATAC (i.e. Cusanovich et al. 2018)}
\usage{
runLSI(mat, nComponents, binarize = FALSE)
}
\arguments{
\item{mat}{raw count matrix from Seurat object in sparse format.}

\item{nComponents}{number of right singular vectors to estimate.}

\item{binarize}{whether to binarize matrix, standardized to FALSE.}
}
\description{
Calculate the Term Frequency - Inverse Document Frequency (TF-IDF) for a feature x cell counts
matrix, then calculate the Singular Value Decomposition of that matrix, which is then used as
input for Seurat's SNN clustering
}
