% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{preprocess_AmbientRNA}
\alias{preprocess_AmbientRNA}
\title{Ambient RNA approximation on Seurat object}
\usage{
preprocess_AmbientRNA(seurat_obj, seed = 1)
}
\description{
Fully runs the ambient RNA approximation protocol through DecontX.
}
