% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{remove_lowRNA}
\alias{remove_lowRNA}
\title{Removes cells with deficient RNA after Ambient RNA removal}
\usage{
remove_lowRNA(seurat_list)
}
\arguments{
\item{seurat_list}{List of Seurat objects}
}
\description{
Removes cells with deficient RNA after Ambient RNA removal
}
